If no md5 then check bytesize

+++ jgi-query.py
@@ -272,7 +272,7 @@ def get_sizes(d, sizes_by_url=None):
     return sizes_by_url
 
 
-def clean_exit(exit_message=None, remove_temp=True):
+def clean_exit(exit_message=None, remove_temp=True, failed=True):
     """
     Perform a sys.exit() while removing temporary files and
     informing the user.
@@ -287,7 +287,7 @@ def clean_exit(exit_message=None, remove_temp=True):
             pass
     for f in to_remove:
        try:
-           os.remove(f)
+           if not failed: os.remove(f) # only remove xml or cookies file when download successfully. when download error, user can check the xml
        except OSError:
            continue
     if exit_message:
@@ -296,7 +296,8 @@ def clean_exit(exit_message=None, remove_temp=True):
         print_message = ""
 
     print("{}Removing temp files and exiting".format(print_message))
-    sys.exit(0)
+    exit_code = 1 if failed else 0
+    sys.exit(exit_code)
     
 
 def extract_file(file_path, keep_compressed=False):
@@ -379,7 +380,7 @@ def print_data(data, org_name, display=True):
     """
     print("\nQUERY RESULTS FOR '{}'\n".format(org_name))
     dict_to_get = {}
-    url_to_validate = {}
+    url_to_validate = defaultdict(dict)
     for query_cat, v in sorted(iter(data.items()),
                                key=lambda k_v: k_v[1]["catID"]):
         print_list = []
@@ -396,13 +397,12 @@ def print_data(data, org_name, display=True):
             for index, i in sorted(items.items()):
                 url = i["url"]
                 dict_to_get[catID][index] = url
-                if url not in url_to_validate: url_to_validate[url] = {}
                 if "md5" in i:
-                    url_to_validate[url]['md5'] = i["md5"]
+                    url_to_validate[url]["md5"] = i["md5"]
                 elif "sizeInBytes" in i:
-                    url_to_validate[url]['sizeInBytes'] = int(i["sizeInBytes"])
+                    url_to_validate[url]["sizeInBytes"] = int(i["sizeInBytes"])
                 elif display is True:
-                    print(f"warn: no md5 or sizeInBytes, so the downloaded file maybe not intact for {org_name}.")
+                    print(f"warn: no md5 or sizeInBytes, so the downloaded file maybe not intact for filename="+i["filename"]+".")
                 print_index = " {}:[{}] ".format(str(catID), str(index))
                 date = fmt_timestamp(i["timestamp"])
                 date_string = '{:02d}/{}'.format(date.tm_mon, date.tm_year)
@@ -435,7 +435,7 @@ def get_user_choice():
     elif choice.lower() in ("q", "quit", "exit"):
         remove_temp = input("Remove index file? (y/n): ")
         remove_temp = remove_temp.lower() in ('y', 'yes', '')
-        clean_exit(remove_temp=remove_temp)
+        clean_exit(remove_temp=remove_temp, failed=False)
     else:
         return choice
 
@@ -585,7 +585,8 @@ def is_broken(filename, min_size_bytes=20, md5_hash=None, sizeInBytes=None):
         not os.path.isfile(filename) or
         os.path.getsize(filename) < min_size_bytes or 
         (is_xml(filename) and not filename.lower().endswith('xml') or
-        not check_md5(filename, md5_hash) or not check_sizeInBytes(filename, sizeInBytes))
+        (md5_hash and not check_md5(filename, md5_hash)) or 
+        (sizeInBytes != None and not check_sizeInBytes(filename, sizeInBytes)) )
     ):
         return True
     else:
@@ -1035,8 +1036,8 @@ LOCAL_XML = False
 if args.load_failed:
     logfile = args.load_failed
     print("Reading URLs from \'{}\'".format(logfile))
-    retry_from_failed(LOGIN_STRING, logfile)
-    clean_exit("All files in log attempted.")
+    downloaded, failed = retry_from_failed(LOGIN_STRING, logfile)
+    clean_exit("All files in log attempted.", failed=failed)
 
 # Get organism name for query
 org_input = args.organism_abbreviation
@@ -1184,7 +1185,7 @@ if INTERACTIVE:
             print('\n'.join(filenames))
             download = input("Continue with download? (y/n/[p]review files): ").lower()
     if download != "y":
-        clean_exit("ABORTING DOWNLOAD")
+        clean_exit("ABORTING DOWNLOAD", failed=False)
 
 downloaded_files, failed_urls = download_list(
     urls_to_get, url_to_validate=url_to_validate, retries=args.retry_n)
@@ -1220,10 +1221,11 @@ if INTERACTIVE:
     keep_temp = input("Keep temporary files ('{}' and 'cookies')? (y/n): "
                     .format(xml_index_filename))
     if keep_temp.lower() not in "y, yes":
-        clean_exit()
+        clean_exit(failed=failed_urls)
     else:
         print("Leaving temporary files intact and exiting.")
 else:
-    clean_exit()
+    clean_exit(failed=failed_urls)
 
-sys.exit(0)
+exit_code = 1 if failed_urls else 0
+sys.exit(exit_code)
